Show command should take release as -r arg
The `show` command accepted a positional release argument but all other
commands accept a release argument using the `-r/--release` option so
this commit changes `show` to do the same for consistency. That frees up
the use of the positional argument so I have chosen to accept a regular
expression given that I frequently do this, e.g:

    $ pystand show -a | grep 3.13.*x86_64.*unknown-linux-gnu.*thread

So now I can just do:

    $ pystand show -a 3.13.*x86_64.*unknown-linux-gnu.*thread

which gives the same output.

Have also added a section about searching using this approach in the
README.