Get latest release tag reliably
- Was getting latest release tag from
https://raw.githubusercontent.com/indygreg/python-build-standalone/latest-release/latest-release.json
but that updates early and points to pre-releases.

- Then used
https://github.com/astral-sh/python-build-standalone/releases.atom which
also was including the pre-release. Even though it was not tagged it did
have a creation timestamp so I was applying a 6 hour delay from
creation before I considered it a release.

- Now am using
https://github.com/astral-sh/python-build-standalone/releases/latest
which normally redirects to the latest release page but I catch the
redirection url and extract the tag. This redirection corresponds
directly to what github flags as "latest" on the releases page which is
ultimately what is wanted.